Improve path config Paths to templates and query definitions were hard-coded. This commit changes that. Instead the paths to be used are defined in "config.toml".
More than one path can be defined. All given paths will be searched (in order). If a path is relative it will be resolved relative to the directory that the "query" package is in.

In order to keep the defaults working, .gitkeep files have been added to the folders: definitions and templates.