Add tasks
Run tasks
The "run_task" script executes a query from a task specified in a json file. It forwards any keyword arguments defined in the json file to the `execute_query` function. By default the results are stored in a temporary file but an alternative path may be given.

A bash script is also added to make it easier to run a task from the command line.

Tasks utility
The "tasks" module in "utils" adds `execute_from_task`. This function will write a json file to the specified task path, wait for the task to be performed, load the data and then delete the temporary files.

This allows someone with access to the tasks and output paths but not to the database to send a request for data retrieval. For this to work the server needs to be set up in a way that:
- tasks folder is monitored for new files
- when a new tasks is detected the `run_task` script is run

The observer script is not part of this package. But it can be set up with for example the watchdog library (https://github.com/gorakhargosh/watchdog).