Add utility query wrapper [New feature under development]
In order to add flexibility to defined queries, a wrapper template was created that will wrap a query and add additional specifications to the query. Currently the user modify the query using keyword arguments when executing the query in the following ways:

- select (list[str]): Select only specified columns.
- where (list[str]): Select only rows that meet where criteria.
- order_by (list[str]): Order by.
- n (int): Fetch only first n records.
- random (bool): Randomize order of rows.

If (and only if) one of these keywords is detected as input to the `get_sql` function then the query will be passed through a parser (utilizing sqlparse: https://github.com/andialbrecht/sqlparse) that will split the query into a body and the final sql statement. The final statement then gets wrapped in the utility template which is what allows for modifying the query on the fly.