Add kwd defaults to config This commit adds functionality so that a user may set defaults in "config.toml" that will be used throughout the application. This is specifically designed to work with kwargs in the context of template rendering.
For this a decorator `add_keyword_defaults` has been added to `utils`. It takes a dictionary that will be used as a base. This dictionary will be passed on to kwargs unless they are overridden by the user.

For the aggregation functionality of the sql wrapper the following defaults have been added to "config.toml":
- label_totals
- label_na