Update `get_paths_from_config` The CONFIG file may contain paths as str, as list or as dict. By default the `get_paths_from_config` function will coerce these types into a list of paths. A `keep_type` flag was added that will ensure that the original type is retained.
This in turn removes the need for the `get_path_from_config` function, as we can now get a Path back from str. Therefore, this function was removed and the "osiris" module now uses `get_paths_from_config` to fetch the location for the credentials.

The docstrings were also edited.