Fix loading from custom `ENV`
You could pass your own `jinja2.Environment` but `get_sql()` would always use `TEMPLATE_LOADER` to try and fetch the template. Now this is changed so that `get_sql()` uses the loader from the `env` that was passed to it. This makes it possible to load queries from custom locations.