Do not relabel if not necessary
If neither `totals` nor `keep_na` is true, then the data does not need to be relabelled (i.e. there is no need to add rows named 'Totals' or '<NA>'). This makes it possible to retain the data type when aggregating in these cases.

On the other hand if we do add totals or null values then we need to convert whatever data type the column is to string. Otherwise the sql statement will fail.